Added more functions
HTML - added table with thead and tbody for events.

CSS - added style for table, tr and th. Changed size of map.

JS - changed map zoom and coordinates for center. Added max and min zoom atributes, also set map atribute draggable to false. Added array of events with their attributes that will be show in added tbody.  Added function that only shows one infoWindow when you click on one marker at the time. Added event for markers when you click on them they show in table events for that specific marker. Discovered new functionst insertRow() and insertCell() for table elements. Also added to event for marker to setCenter to this koordinate and to zoom on that marker.

???Need to find way to when change marker to zoom on other marker position???